Move Dask Client configuration to Component class and use multi-GPU in `embed_images` component (#852)
It doesn't really make sense to let the user configure the Dask client
used by a component, since this is specific to the component
implementation. I therefore moved it to the Component class with a
default implementation of the Dask `LocalCluster`. Note that due to a
bug the `LocalCluster` was not yet used as default. This therefore
changes the default for all components, which were using the threaded
scheduler before.

One downside to the current implementation is that we require the
component developer to call `super().__init__()` in their `__init__`
method. This is best practice though and highlighted by IDEs.