Infer consume operation if not present in dataset interface (#859)
Basic implementation, I still have to add tests. I wanted to get some
feedback first.

- Added a method to the lightweight components to generate a
`ComponentSpec` based on the attributes.
- Added a method in the pipeline to infer the consumption based on the
`ComponentSpec`.
In cases where a user hasn't specified a `consume` in the pipeline
operations, we now infer this. If a component spec contains a `consumes`
section and `additionalProperties` are set to true, we load all columns.
If `additionalProperties` is set to false, we limit the columns defined
in the component spec.

Fix #836