Change dask_client to general setup method (#861)
This PR changes the `dask_client` method introduced in #852 into a
general `setup` method.

This method differs from `__init__` in that it allows users to return a
state, which is passed into the `teardown` method by Fondant. This is
necessary since the Dask client is not pickleable, and setting it as an
instance attribute leads to issues when executing the `transform` method
across processes (as is the case in the `PandasTransformComponent`).