Add pre- and post-build script to work around Poetry bug (#868)
The new packaging strategy introduced in #849 doesn't work correctly for
wheels due to a bug in poetry
(https://github.com/python-poetry/poetry/issues/8994), which breaks
installing both from Github and PyPi. This PR introduces a workaround.

Since the bug is related to the `src` directory, this workaround adds a
`pre-build/sh` script which unpacks the `src` directory before building,
and a `post-build.sh` script which moves the package back into the `src`
directory afterwards.

This leads to the following result:
- Installing from Github leads to an installation including all
component files
- Installing locally without running the build script leads to an
installation including all component files
- Installing from PyPi leads to an installation with only the
`fondant_component.yaml` files for each component

In the ideal case, we get the same result for all installation methods,
but this already at least leads to a working result for each method.