Move convert-string into component setup method (#871)
Dask config needs to be set before starting the `LocalCluster`. I first
put it into the `DaskDataIO` since it's central to the working of
Fondant that this is set correctly, and I didn't want this to be
overwritten. That's too late in our flow though. The other place we
could put it is in the executor, but we currently try to keep that
Dask-agnostic.