Add string as fallback index type when writing data (#904)
`pa.from_numpy_dtype` fails when passing an object dtype, which Dask
uses when the dtype is not explicitly known. I propose to fall back on
`string` if this is the case. This will probably be correct for the
index column in most cases when the `dtype` is unknown, and non-breaking
in most other cases.

---------

Co-authored-by: Matthias Richter <matthias.r1092@gmail.com>