Skip transformation of partition if partition is empty (#908)
In the case that a `PandasTransformComponent` receives an empty dask
partition, the execution fails. There could be several reasons why a
partition is empty, and it may depend on a custom implementation. We
should catch this case and not halt the component's execution.
This PR adds a check to see if the `wrapped_transform` receives an empty
partition. If it does, we skip the execution and return the received
dataframe.

---------

Co-authored-by: Robbe Sneyders <robbe.sneyders@gmail.com>