Fixing cicd (#929)
dask <= 2024.4.0 is incompatible with python 3.11.9. Github updated
there runners and our cicd pipelines were failing.
A fix was released as part of dask 2024.4.1, however we can not use this
version since there are dependency conflicts with dask-cuda.

<s>Limited our python version to <3.11.9</s>

**Update 16/04/24:** 
We encountered `TypeError "<" not supported between instance of 'int'
and 'str'`. This is related to sorting operation part of `dask-expr`.
dask-expr is part of `dask>=2024.3.0`. Dask supports a specific
configuration to disable the computation graph optimizations and the
usage of dask-expr. However, limit the dask version to `dask<2024.3.0`
would avoid further issues since we have tested these version before.